refactor(modules): extract "mixin" and "argsToParams" static methods from main module
