+++ src/staticMethods/index.js
@@ -1,0 +1,2 @@
+export * from './argsToParams'
+export * from './mixin'
